feat: Introduce base service classes and refactor user service for improved structure
