+++ aura/users/services/user/service.py
@@ -1,15 +1,7 @@
-from __future__ import annotations
-
-import threading
-import typing
-from abc import ABC
 from abc import abstractmethod
-from typing import Self
 
+from core.services.base import Service
 from typing_extensions import TypedDict
-
-if typing.TYPE_CHECKING:
-    from aura.users.models import User
 
 
 class UserFilterArgs(TypedDict, total=False):
@@ -18,9 +10,6 @@ class UserFilterArgs(TypedDict, total=False):
 
     is_active: bool
     """Whether the user needs to be active"""
-
-    organization_id: int
-    """Organization to check membership in"""
 
     emails: list[str]
     """list of emails to match with"""
@@ -35,64 +24,6 @@ class UserFilterArgs(TypedDict, total=False):
     """The type of MFA authenticator you want to query by"""
 
 
-class DelegatingService:
-    def __init__(
-        self,
-        base_service_cls,
-        constructor,
-        signatures,
-    ) -> None:
-        self._constructor = constructor
-        self._singleton = {}
-        self._lock = threading.RLock()
-        self._base_service_cls = base_service_cls
-
-    def __getattr__(self, item: str):
-        key = self._base_service_cls.key
-
-        try:
-            # fast path: object already built
-            impl = self._singleton[key]
-        except KeyError:
-            # slow path: only lock when building the object
-            with self._lock:
-                # another thread may have won the race to build the object
-                try:
-                    impl = self._singleton[key]
-                except KeyError:
-                    impl = self._singleton[key] = self._constructor()
-
-        return getattr(impl, item)
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}({self._base_service_cls.__name__})"
-
-
-class Service(ABC):
-    @classmethod
-    @abstractmethod
-    def get_local_implementation(cls) -> Service:
-        """Return a service object that runs locally.
-
-        The returned service object is (generally) the database-backed instance that
-        is called when we either receive a remote call from outside, or want to call
-        it within the same silo.
-
-        A base service class generally should override this class method, making a
-        forward reference to its own database-backed subclass.
-        """
-
-        raise NotImplementedError
-
-    @classmethod
-    def create_delegation(cls, use_test_client: bool | None = None) -> Self:
-        """Instantiate a base service class for the current mode."""
-        constructor = cls.get_local_implementation()
-        service = DelegatingService(cls, constructor)
-        # this returns a proxy which simulates the given class
-        return service  # noqa: RET504
-
-
 class UserService(Service):
     key = "user"
 
@@ -103,10 +34,10 @@ class UserService(Service):
         return DatabaseBackedUserService()
 
     @abstractmethod
-    def get_many(self, *, filter: UserFilterArgs) -> list[User]:
+    def get_many(self, *, filter: UserFilterArgs):
         pass
 
-    def get_user(self, user_id: int) -> User | None:
+    def get_user(self, user_id: int):
         """
         Get a single user by id
 
@@ -117,7 +48,7 @@ class UserService(Service):
         return get_user(user_id)
 
 
-def get_user(user_id: int) -> User | None:
+def get_user(user_id: int):
     users = user_service.get_many(filter={"user_ids": [user_id]})
     if len(users) > 0:
         return users[0]
